requirements.txt
fastapi
uvicorn
cassandra-driver
redis
